workflow/requirements.txt
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
textual>=0.61.0
# Gmail sending (Agent 5)
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
backoff>=2.2.1
